added color picker to fragment
