+++ app/src/main/java/cheetatech/com/colorhub/ColorPickerActivity.java
@@ -12,6 +12,8 @@ import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
 import android.support.design.widget.TabLayout;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
@@ -29,9 +31,12 @@ import android.widget.Toast;
 
 import org.json.JSONObject;
 
+import java.util.List;
+
 import cheetatech.com.colorhub.adapters.ViewPagerAdapter;
 import cheetatech.com.colorhub.defines.BoardEditor;
 import cheetatech.com.colorhub.defines.ColorItem;
+import cheetatech.com.colorhub.listeners.IOnFocusListenable;
 import layout.ColorPicker1;
 import layout.ColorPicker2;
 import layout.FlatColorFragment;
@@ -40,32 +45,10 @@ import layout.MaterialColorFragment;
 import layout.MetroColorFragment;
 import layout.SocialColorFragment;
 
-public class ColorPickerActivity extends AppCompatActivity implements TabLayout.OnTabSelectedListener ,SeekBar.OnSeekBarChangeListener, View.OnClickListener {
-
-    View colorView;
-    View[] views = null;
-    int[] viewIds = new int[]{R.id.color1,R.id.color2,R.id.color3,R.id.color4,R.id.color5};
-    SeekBar redSeekBar, greenSeekBar, blueSeekBar,opacitySeekBar;
-    TextView redToolTip, greenToolTip, blueToolTip,opacityToolTip;
-    Button buttonSelector;
-    ClipboardManager clipBoard;
-    ClipData clip;
-    Window window;
-    Display display;
-    int red, green, blue, seekBarLeft,opacity;
-    Rect thumbRect;
+public class ColorPickerActivity extends AppCompatActivity implements TabLayout.OnTabSelectedListener {
     private TabLayout tabLayout = null;
-
     private ViewPager viewPager = null;
-
-    ColorItem[] colors = new ColorItem[]{
-            new ColorItem(0,0,0,0),
-            new ColorItem(0,0,0,0),
-            new ColorItem(0,0,0,0),
-            new ColorItem(0,0,0,0),
-            new ColorItem(0,0,0,0),
-    };
-
+    private ColorPicker1 ColorPicker_1 = null;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -91,304 +74,43 @@ public class ColorPickerActivity extends AppCompatActivity implements TabLayout.
 
         tabLayout.setOnTabSelectedListener(this);
 
-
-        display = ((WindowManager) this.getSystemService(Context.WINDOW_SERVICE)).getDefaultDisplay();
-
-        red = green = blue = opacity = 255;
-
-        clipBoard = (ClipboardManager)getSystemService(Context.CLIPBOARD_SERVICE);
-        colorView = findViewById(R.id.colorView);
-        window = getWindow();
-
-        redSeekBar = (SeekBar)findViewById(R.id.redSeekBar);
-        greenSeekBar = (SeekBar)findViewById(R.id.greenSeekBar);
-        blueSeekBar = (SeekBar)findViewById(R.id.blueSeekBar);
-
-        opacitySeekBar = (SeekBar)findViewById(R.id.opacitySeekBar);
-
-        seekBarLeft = redSeekBar.getPaddingLeft();
-
-        redToolTip = (TextView)findViewById(R.id.redToolTip);
-        greenToolTip = (TextView)findViewById(R.id.greenToolTip);
-        blueToolTip = (TextView)findViewById(R.id.blueToolTip);
-
-        opacityToolTip = (TextView)findViewById(R.id.opacityToolTip);
-
-        buttonSelector = (Button)findViewById(R.id.buttonSelector);
-        buttonSelector.setOnClickListener(this);
-
-        redSeekBar.setOnSeekBarChangeListener(this);
-        greenSeekBar.setOnSeekBarChangeListener(this);
-        blueSeekBar.setOnSeekBarChangeListener(this);
-        opacitySeekBar.setOnSeekBarChangeListener(this);
-
-
-        redSeekBar.setProgress(red);
-        greenSeekBar.setProgress(green);
-        blueSeekBar.setProgress(blue);
-        opacitySeekBar.setProgress(opacity);
-
-        colorView.setBackgroundColor(Color.argb(opacity,red, green, blue));
-
-        buttonSelector.setText(String.format("#%02x%02x%02x%02x", red, green, blue,opacity));
-
-        views = new View[5];
-        for (int i = 0;i<viewIds.length;i++)
-        {
-            views[i] = (View)findViewById(viewIds[i]);
-            views[i].setOnClickListener(this);
-        }
-
-        for(int j = 0;j<colors.length;j++)
-            views[j].setBackgroundColor(Color.argb(colors[j].getOpacity(),colors[j].getRed(),colors[j].getGreen(),colors[j].getBlue()));
-
     }
-
-
-
     public  void setUpViewPager(ViewPager viewPager)
     {
+        if(ColorPicker_1 == null)
+            ColorPicker_1 = new ColorPicker1();
         ViewPagerAdapter adapter = new ViewPagerAdapter(getSupportFragmentManager());
-        adapter.addFragment(new ColorPicker1(),"ColorPicker1");
+        adapter.addFragment(ColorPicker_1,"ColorPicker1");
         adapter.addFragment(new ColorPicker2(),"ColorPicker2");
         viewPager.setAdapter(adapter);
     }
-
-
-    @Override
-    public void onWindowFocusChanged(boolean hasFocus) {
-
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-            thumbRect = redSeekBar.getThumb().getBounds();
-        }
-
-        redToolTip.setX(seekBarLeft + thumbRect.left);
-        if (red<10)
-            redToolTip.setText("  "+red);
-        else if (red<100)
-            redToolTip.setText(" "+red);
-        else
-            redToolTip.setText(red+"");
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-            thumbRect = greenSeekBar.getThumb().getBounds();
-        }
-
-        greenToolTip.setX(seekBarLeft + thumbRect.left);
-        if (green<10)
-            greenToolTip.setText("  "+green);
-        else if (red<100)
-            greenToolTip.setText(" "+green);
-        else
-            greenToolTip.setText(green+"");
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-            thumbRect = blueSeekBar.getThumb().getBounds();
-        }
-
-        blueToolTip.setX(seekBarLeft + thumbRect.left);
-        if (blue<10)
-            blueToolTip.setText("  "+blue);
-        else if (blue<100)
-            blueToolTip.setText(" "+blue);
-        else
-            blueToolTip.setText(blue+"");
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-            thumbRect = opacitySeekBar.getThumb().getBounds();
-        }
-
-        opacityToolTip.setX(seekBarLeft + thumbRect.left);
-        if (opacity<10)
-            opacityToolTip.setText("  "+opacity);
-        else if (opacity<100)
-            opacityToolTip.setText(" "+opacity);
-        else
-            opacityToolTip.setText(opacity+"");
-
-    }
-
-    @TargetApi(Build.VERSION_CODES.JELLY_BEAN)
-    @Override
-    public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
-
-        if (seekBar.getId() == R.id.redSeekBar) {
-
-            red = progress;
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-                thumbRect = seekBar.getThumb().getBounds();
-            }
-
-            redToolTip.setX(seekBarLeft + thumbRect.left);
-
-            if (progress<10)
-                redToolTip.setText("  " + red);
-            else if (progress<100)
-                redToolTip.setText(" "+red);
-            else
-                redToolTip.setText(red+"");
-
-        }
-        else if (seekBar.getId() == R.id.greenSeekBar) {
-
-            green = progress;
-            thumbRect = seekBar.getThumb().getBounds();
-
-            greenToolTip.setX(seekBar.getPaddingLeft()+thumbRect.left);
-            if (progress<10)
-                greenToolTip.setText("  "+green);
-            else if (progress<100)
-                greenToolTip.setText(" "+green);
-            else
-                greenToolTip.setText(green+"");
-
-        }
-        else if (seekBar.getId() == R.id.blueSeekBar) {
-
-            blue = progress;
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-                thumbRect = seekBar.getThumb().getBounds();
-            }
-
-            blueToolTip.setX(seekBarLeft + thumbRect.left);
-            if (progress<10)
-                blueToolTip.setText("  "+blue);
-            else if (progress<100)
-                blueToolTip.setText(" "+blue);
-            else
-                blueToolTip.setText(blue+"");
-
-        }
-        else if (seekBar.getId() == R.id.opacitySeekBar) {
-
-            opacity = progress;
-
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-                thumbRect = seekBar.getThumb().getBounds();
-            }
-            opacityToolTip.setX(seekBarLeft + thumbRect.left);
-            if (progress<10)
-                opacityToolTip.setText("  "+opacity);
-            else if (progress<100)
-                opacityToolTip.setText(" "+opacity);
-            else
-                opacityToolTip.setText(opacity+"");
-
-        }
-
-        colorView.setBackgroundColor(Color.argb(opacity,red, green, blue));
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-
-            if (display.getRotation() == Surface.ROTATION_0)
-                window.setStatusBarColor(Color.argb(opacity,red, green, blue));
-        }
-        buttonSelector.setText(String.format("#%02x%02x%02x%02x", red, green, blue,opacity));
-    }
-
-    private void addToView(ColorItem colorItem) {
-
-        if(views != null) {
-        ColorItem[] refColors = new ColorItem[6];
-        int i;
-        for(i = 0;i<colors.length;i++)
-            refColors[i] = colors[i];
-            views[0].setBackgroundColor(Color.BLUE);
-            views[0].setBackgroundColor(Color.argb(colorItem.getOpacity(), colorItem.getRed(), colorItem.getGreen(), colorItem.getBlue()));
-            for (i = 0; i < colors.length - 1; i++)
-                views[i + 1].setBackgroundColor(Color.argb(refColors[i].getOpacity(), refColors[i].getRed(), refColors[i].getGreen(), refColors[i].getBlue()));
-
-            colors[0] = new ColorItem(colorItem.getRed(), colorItem.getGreen(), colorItem.getBlue(), colorItem.getOpacity());
-
-            for (i = 0; i < colors.length - 1; i++)
-                colors[i + 1] = new ColorItem(refColors[i].getRed(), refColors[i].getGreen(), refColors[i].getBlue(), refColors[i].getOpacity());
-        }
-    }
-
-
-    private void setProgressBar(int r,int g,int b,int o)
-    {
-
-        redSeekBar.setProgress(r);
-        greenSeekBar.setProgress(g);
-        blueSeekBar.setProgress(b);
-        opacitySeekBar.setProgress(o);
-    }
-
-    @Override
-    public void onStopTrackingTouch(SeekBar seekBar) {
-
-    }
-
-    @Override
-    public void onStartTrackingTouch(SeekBar seekBar) {
-
-    }
-
-
-
     @Override
     protected void onDestroy() {
         super.onDestroy();
     }
 
     @Override
-    public void onClick(View view) {
-        switch (view.getId())
+    public void onWindowFocusChanged(boolean hasFocus)
+    {
+        super.onWindowFocusChanged(hasFocus);
+        Fragment currentFragment = getVisibleFragment();
+        if(currentFragment != null)
         {
-            case R.id.buttonSelector:
-                addToView(new ColorItem(red,green,blue,opacity));
-                BoardEditor.getInstance().copyToClipBoard(buttonSelector.getText().toString());
-                Toast.makeText(BoardEditor.getInstance().getContext(), "Color " + buttonSelector.getText().toString() +
-                        " copied to clipboard...", Toast.LENGTH_SHORT).show();
-                break;
-            case R.id.color1:
-
-                BoardEditor.getInstance().copyToClipBoard(colors[0].toString());
-                Toast.makeText(BoardEditor.getInstance().getContext(), "Color " + colors[0].toString() +
-                        " copied to clipboard...", Toast.LENGTH_SHORT).show();
-                red = colors[0].getRed(); green = colors[0].getGreen(); blue = colors[0].getBlue(); opacity = colors[0].getOpacity();
-                setProgressBar(red,green,blue,opacity);
-                onWindowFocusChanged(true);
-                break;
-            case R.id.color2:
-
-                BoardEditor.getInstance().copyToClipBoard(colors[1].toString());
-                Toast.makeText(BoardEditor.getInstance().getContext(), "Color " + colors[1].toString() +
-                        " copied to clipboard...", Toast.LENGTH_SHORT).show();
-                red = colors[1].getRed(); green = colors[1].getGreen(); blue = colors[1].getBlue(); opacity = colors[1].getOpacity();
-                setProgressBar(red,green,blue,opacity);
-                onWindowFocusChanged(true);
-                break;
-            case R.id.color3:
-
-                BoardEditor.getInstance().copyToClipBoard(colors[2].toString());
-                Toast.makeText(BoardEditor.getInstance().getContext(), "Color " + colors[2].toString() +
-                        " copied to clipboard...", Toast.LENGTH_SHORT).show();
-                red = colors[2].getRed(); green = colors[2].getGreen(); blue = colors[2].getBlue(); opacity = colors[2].getOpacity();
-                setProgressBar(red,green,blue,opacity);
-                onWindowFocusChanged(true);
-                break;
-            case R.id.color4:
-
-                BoardEditor.getInstance().copyToClipBoard(colors[3].toString());
-                Toast.makeText(BoardEditor.getInstance().getContext(), "Color " + colors[3].toString() +
-                        " copied to clipboard...", Toast.LENGTH_SHORT).show();
-                red = colors[3].getRed(); green = colors[3].getGreen(); blue = colors[3].getBlue(); opacity = colors[3].getOpacity();
-                setProgressBar(red,green,blue,opacity);
-                onWindowFocusChanged(true);
-                break;
-            case R.id.color5:
-
-                BoardEditor.getInstance().copyToClipBoard(colors[4].toString());
-                Toast.makeText(BoardEditor.getInstance().getContext(), "Color " + colors[4].toString() +
-                        " copied to clipboard...", Toast.LENGTH_SHORT).show();
-                red = colors[4].getRed(); green = colors[4].getGreen(); blue = colors[4].getBlue(); opacity = colors[4].getOpacity();
-                setProgressBar(red,green,blue,opacity);
-                onWindowFocusChanged(true);
-                break;
+            if(currentFragment instanceof  IOnFocusListenable)
+                ((IOnFocusListenable)currentFragment).onWindowFocusChanged(hasFocus);
         }
+
+    }
+    public Fragment getVisibleFragment(){
+        FragmentManager fragmentManager = ColorPickerActivity.this.getSupportFragmentManager();
+        List<Fragment> fragments = fragmentManager.getFragments();
+        if(fragments != null){
+            for(Fragment fragment : fragments){
+                if(fragment != null && fragment.isVisible())
+                    return fragment;
+            }
+        }
+        return null;
     }
 
     @Override
