+++ app/src/main/java/cheetatech/com/colorhub/listeners/IOnFocusListenable.java
@@ -1,0 +1,6 @@
+package cheetatech.com.colorhub.listeners;
+
+
+public interface IOnFocusListenable {
+    public  void onWindowFocusChanged(boolean hasFocus);
+}
